# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Gestor De Pedidos",
        "version": "1.0.0"
    },
    "paths": {
        "/api/v1/login/": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Return a customer logged with jwt token",
                "description": "Return a customer logged with jwt token",
                "operationId": "18b12149048e76fd590a0d8f6939e087",
                "parameters": [
                    {
                        "name": "email",
                        "in": "query",
                        "description": "Customer Email",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "password",
                        "in": "query",
                        "description": "Customer Password",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Customer Logged"
                    }
                }
            }
        },
        "/api/v1/customers": {
            "get": {
                "tags": [
                    "Customers"
                ],
                "summary": "Returns a list of customers",
                "description": "Returns a object of all customers",
                "operationId": "fddf8e3800951c00bc5fbd50026d7e6a",
                "responses": {
                    "200": {
                        "description": "list of Customers"
                    }
                },
                "security": [
                    {
                        "token": []
                    }
                ]
            }
        },
        "/api/v1/customer/{id}": {
            "get": {
                "tags": [
                    "Customers"
                ],
                "summary": "Returns a one customer",
                "description": "Returns a object of customer",
                "operationId": "aba9fdb513d2739249feff401c6798bf",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Customer Id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "One Customer"
                    }
                },
                "security": [
                    {
                        "token": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Customers"
                ],
                "summary": "Return a customer updated",
                "description": "Return a customer updated",
                "operationId": "e7f34a0d2fa8eef45d1e54b99b923c8c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Customer Id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "name",
                        "in": "query",
                        "description": "Customer Name",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "email",
                        "in": "query",
                        "description": "Customer Email",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "update Customer"
                    }
                },
                "security": [
                    {
                        "token": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Customers"
                ],
                "summary": "Return a customer deleted",
                "description": "Return a customer deleted",
                "operationId": "9dfc788762aab040f73d666a480628ac",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Customer Id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "delete Customer"
                    }
                },
                "security": [
                    {
                        "token": []
                    }
                ]
            }
        },
        "/api/v1/customer/store": {
            "post": {
                "tags": [
                    "Customers"
                ],
                "summary": "Return a customer created",
                "description": "Return a customer created",
                "operationId": "4ee0a95ab90a862d9150ca7f54c83964",
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "description": "Customer Name",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "email",
                        "in": "query",
                        "description": "Customer Email",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "password",
                        "in": "query",
                        "description": "Customer Password",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Create Customer"
                    }
                }
            }
        },
        "/api/v1/products": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "Returns a list of products",
                "description": "Returns a object of all products",
                "operationId": "072fd15d3be55577bb08c19d19f1e75c",
                "responses": {
                    "200": {
                        "description": "list of products"
                    }
                },
                "security": [
                    {
                        "token": []
                    }
                ]
            }
        },
        "/api/v1/product/{$id}": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "Returns a one product",
                "description": "Returns a one product",
                "operationId": "5dd7b61a41c37b6cbb9bfc9c65b399f0",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Product Id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "one product"
                    }
                },
                "security": [
                    {
                        "token": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Products"
                ],
                "summary": "Returns a product deleted",
                "description": "Returns a product deleted",
                "operationId": "310d8a90a3352a7b74638bbfd7c34c5e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Product Id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "return a product deleted"
                    }
                },
                "security": [
                    {
                        "token": []
                    }
                ]
            }
        },
        "/api/v1/product/{id}": {
            "put": {
                "tags": [
                    "Products"
                ],
                "summary": "Return a product updated",
                "description": "Return a product updated",
                "operationId": "4c7982ff42a6c504de8ccf7ecf205d30",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "`Product Id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "name",
                        "in": "query",
                        "description": "Product Name",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "quantity",
                        "in": "query",
                        "description": "Product Quantity",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "description",
                        "in": "query",
                        "description": "Product Description",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "price",
                        "in": "query",
                        "description": "Product Price",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "status",
                        "in": "query",
                        "description": "Product Status",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "update Product"
                    }
                },
                "security": [
                    {
                        "token": []
                    }
                ]
            }
        },
        "/api/v1/product": {
            "post": {
                "tags": [
                    "Products"
                ],
                "summary": "Return a product created",
                "description": "Return a product created",
                "operationId": "11b374caabe03b71870182cdbef8275d",
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "description": "Product Name",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "quantity",
                        "in": "query",
                        "description": "Product Quantity",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "description",
                        "in": "query",
                        "description": "Product Description",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "price",
                        "in": "query",
                        "description": "Product Price",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "status",
                        "in": "query",
                        "description": "Product Status",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Create Product"
                    }
                },
                "security": [
                    {
                        "token": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "Bearer": {
                "type": "apiKey",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}